+++ src/services/dispatch.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from "@angular/core";
 import {Store} from "@ngrx/store";
 
-import {GenericResource, GenericAction} from "../resource";
+import {GenericAction, GenericActionTypesVariants, GenericResource} from "../resource";
 import {createAction, IPayload, IResourceActionOptions} from "../utils";
 
 @Injectable()
@@ -11,16 +11,17 @@ export class DispatchService {
     }
 
     dispatch<resource extends GenericResource>(resource: new () => resource,
-                                                action: GenericAction,
-                                                payload?: IPayload): void {
-        const actionInstance = createAction(resource, action, payload);
+                                               action: GenericAction,
+                                               payload?: IPayload,
+                                               variant?: GenericActionTypesVariants): void {
+        const actionInstance = createAction(resource, action, payload, undefined, variant);
         this.store.dispatch(actionInstance);
     }
 
     dispatchComplex<resource extends GenericResource, P>(resource: new () => resource,
-                                                       action: GenericAction,
-                                                       payload?: IPayload,
-                                                       options?: IResourceActionOptions): void {
+                                                         action: GenericAction,
+                                                         payload?: IPayload,
+                                                         options?: IResourceActionOptions): void {
         const actionInstance = createAction(resource, action, payload, options);
         this.store.dispatch(actionInstance);
     }